refactor: update imports to new structure and migrate to better-auth provider
