+++ core/src/zap/api/errors.ts
